// AzureParallelDeploy/AzureParallelDeploy/AzureParallelDeployV3/src/Interfaces.ts
export enum AppType {
	WebApp = "WebApp",
	FunctionApp = "FunctionApp"
}

export interface Settings {
	appType: AppType;
	resourceGroup: string;
	appNameFormat: string;
	appSourceFormat: string;
	appSourceBasePath: string;
	slotName: string;
	subscriptionId: string;
	syncFunctionTriggers: boolean;
}